# prototype/replit/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Crop = (typeof Crop)[keyof typeof Crop];

export const Crop = {
  Rice: "Rice",
  Wheat: "Wheat",
  Maize: "Maize",
  Cotton: "Cotton",
  Sugarcane: "Sugarcane",
  Soybean: "Soybean",
  Groundnut: "Groundnut",
  Tomato: "Tomato",
  Potato: "Potato",
  Onion: "Onion",
} as const;

export interface CreateScanInput {
  crop: Crop;
  /**
   * @maxItems 20
   * @items.minLength 1
   * @items.maxLength 200
   */
  symptoms?: string[];
  /**
   * @minLength 1
   * @maxLength 100
   */
  affectedPart?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  growthStage?: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  affectedAreaPercentage?: number;
  nearbyPlantsAffected?: boolean;
  /** @maxLength 2000 */
  notes?: string;
  /**
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
}

export type ScanStatus = (typeof ScanStatus)[keyof typeof ScanStatus];

export const ScanStatus = {
  pending: "pending",
  analyzing: "analyzing",
  completed: "completed",
  failed: "failed",
} as const;

export type Scan = CreateScanInput & {
  id: string;
  status: ScanStatus;
  /** @nullable */
  imagePath: string | null;
  createdAt: string;
  updatedAt: string;
};

export type ErrorResponseErrorDetailsItem = { [key: string]: unknown };

export type ErrorResponseError = {
  code: string;
  message: string;
  details?: ErrorResponseErrorDetailsItem[];
};

export interface ErrorResponse {
  error: ErrorResponseError;
}