# prototype/replit/lib/api-zod/src/generated/types/createScanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Crop } from './crop';

export interface CreateScanInput {
  crop: Crop;
  /**
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 200
     */
  symptoms?: string[];
  /**
     * @minLength 1
     * @maxLength 100
     */
  affectedPart?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  growthStage?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  affectedAreaPercentage?: number;
  nearbyPlantsAffected?: boolean;
  /** @maxLength 2000 */
  notes?: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
}