# prototype/replit/lib/api-zod/src/generated/types/crop.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Crop = typeof Crop[keyof typeof Crop];


export const Crop = {
  Rice: 'Rice',
  Wheat: 'Wheat',
  Maize: 'Maize',
  Cotton: 'Cotton',
  Sugarcane: 'Sugarcane',
  Soybean: 'Soybean',
  Groundnut: 'Groundnut',
  Tomato: 'Tomato',
  Potato: 'Potato',
  Onion: 'Onion',
} as const;