# prototype/replit/lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseError } from './errorResponseError';

export interface ErrorResponse {
  error: ErrorResponseError;
}