# prototype/replit/lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScanInput } from './createScanInput';
import type { ScanStatus } from './scanStatus';

export type Scan = CreateScanInput & ({
  id: string;
  status: ScanStatus;
  /** @nullable */
  imagePath: string | null;
  createdAt: Date;
  updatedAt: Date;
});